Cards controller edit
Cards controller now inherits from protected controller

Tested all curl scripts for Card and I am now denied token
access, which is a good thing. Now, a current user has to be signed in
to their account to have access to their own cards.

Aside from adding a full README file the API for this project
appears to be complete.